+++ .nycrc.json
@@ -15,6 +15,7 @@
     "src/agents/org-detector/agent.js",
     "src/agents/org-detector/instructions.js",
     "src/controllers/demo.js",
-    "src/controllers/llmo/llmo.js"
+    "src/controllers/llmo/llmo.js",
+    "src/controllers/llmo/brand-presence/*"
   ]
 }
