+++ src/controllers/llmo/llmo.js
@@ -44,6 +44,10 @@ import {
 } from './llmo-onboarding.js';
 import { queryLlmoFiles } from './llmo-query-handler.js';
 import { updateModifiedByDetails } from './llmo-config-metadata.js';
+import {
+  getBrandPresenceFilters as getBrandPresenceFiltersImpl,
+  exampleEndpoint as exampleEndpointImpl,
+} from './brand-presence/index.js';
 
 const { readConfig, writeConfig } = llmo;
 const { llmoConfig: llmoConfigSchema } = schemas;
@@ -944,217 +948,17 @@ function LlmoController(ctx) {
     }
   };
 
-  // Handles requests to the LLMO Aurora endpoint
-  // Tests Aurora PostgreSQL database connectivity and retrieves audit statistics
-  const getLlmoAurora = async (context) => {
-    const {
-      log, env, aurora,
-    } = context;
-    const { siteId } = context.params;
-    const startTime = Date.now();
+  // Wrapper for brand presence filters endpoint
+  const getBrandPresenceFilters = async (context) => getBrandPresenceFiltersImpl(
+    context,
+    getSiteAndValidateLlmo,
+  );
 
-    log.info(`[LLMO-AURORA] Starting request for siteId: ${siteId}`);
-
-    try {
-      // Validate LLMO access
-      log.info(`[LLMO-AURORA] Validating LLMO access for siteId: ${siteId}`);
-      const validationStart = Date.now();
-      await getSiteAndValidateLlmo(context);
-      const validationDuration = Date.now() - validationStart;
-      log.info(`[LLMO-AURORA] LLMO access validation completed for siteId: ${siteId} - duration: ${validationDuration}ms`);
-
-      // Test Aurora database connectivity
-      let dbStats = null;
-      let dbTestDuration = 0;
-      if (aurora && env.ENABLE_AURORA_QUERIES) {
-        try {
-          log.info(`[LLMO-AURORA] Testing Aurora database connectivity for siteId: ${siteId}`);
-          const dbTestStart = Date.now();
-
-          // Test 1: Simple connectivity test
-          const connected = await aurora.testConnection();
-
-          // Test 5: Query brand presence data
-          const brandPresenceData = await aurora.query(
-            `SELECT id, site_id, date, model, category, prompt, region, url, sources, citations, mentions FROM public.brand_presence
-             WHERE date = $1 AND category = $2`,
-            ['2025-11-24', 'Adobe'],
-          );
-
-          // Test 6: Count citations where citations = true
-          const citationCount = await aurora.queryOne(
-            `SELECT COUNT(*) as total_citations
-             FROM public.brand_presence
-             WHERE date = $1 AND category = $2 AND citations = true`,
-            ['2025-11-24', 'Adobe'],
-          );
-
-          dbTestDuration = Date.now() - dbTestStart;
-
-          log.info(`[LLMO-AURORA] Brand presence data: ${JSON.stringify(brandPresenceData)}`);
-          dbStats = {
-            connected,
-            brandPresence: {
-              data: brandPresenceData,
-              totalRecords: brandPresenceData.length,
-              totalCitations: citationCount ? parseInt(citationCount.total_citations, 10) : 0,
-            },
-            poolStats: aurora.getPoolStats(),
-          };
-
-          log.info(`[LLMO-AURORA] Aurora database test completed for siteId: ${siteId} - duration: ${dbTestDuration}ms, connected: ${connected}, audits found: ${dbStats.totalAudits}, brand presence records: ${brandPresenceData.length}, total citations: ${dbStats.brandPresence.totalCitations}`);
-        } catch (dbError) {
-          log.warn(`[LLMO-AURORA] Aurora database test failed for siteId: ${siteId} - error: ${dbError.message}`);
-          dbStats = {
-            connected: false,
-            error: dbError.message,
-          };
-        }
-      } else {
-        log.info(`[LLMO-AURORA] Aurora database not configured or disabled for siteId: ${siteId}`);
-      }
-
-      const totalDuration = Date.now() - startTime;
-      log.info(`[LLMO-AURORA] Request completed for siteId: ${siteId} - total duration: ${totalDuration}ms`);
-
-      return ok({
-        siteId,
-        auroraStats: dbStats,
-        performance: {
-          totalDuration,
-          dbTestDuration,
-          validationDuration: validationDuration || 0,
-        },
-      });
-    } catch (error) {
-      const totalDuration = Date.now() - startTime;
-      log.error(`[LLMO-AURORA] Request failed for siteId: ${siteId} - duration: ${totalDuration}ms, error: ${error.message}, stack: ${error.stack}`);
-      return badRequest(error.message);
-    }
-  };
-
-  // Handles requests to get distinct filter values from brand_presence table
-  // Returns available values for each filter dimension to populate UI dropdowns
-  const getBrandPresenceFilters = async (context) => {
-    const {
-      log, env, aurora,
-    } = context;
-    const { siteId } = context.params;
-    const startTime = Date.now();
-
-    log.info(`[BRAND-PRESENCE-FILTERS] Starting request for siteId: ${siteId}`);
-
-    try {
-      // Validate LLMO access
-      log.info(`[BRAND-PRESENCE-FILTERS] Validating LLMO access for siteId: ${siteId}`);
-      const validationStart = Date.now();
-      await getSiteAndValidateLlmo(context);
-      const validationDuration = Date.now() - validationStart;
-      log.info(`[BRAND-PRESENCE-FILTERS] LLMO access validation completed for siteId: ${siteId} - duration: ${validationDuration}ms`);
-
-      // Check if Aurora is configured and enabled
-      if (!aurora || !env.ENABLE_AURORA_QUERIES) {
-        log.warn(`[BRAND-PRESENCE-FILTERS] Aurora database not configured or disabled for siteId: ${siteId}`);
-        return badRequest('Aurora database is not configured or queries are not enabled');
-      }
-
-      let filterValues = null;
-      let queryDuration = 0;
-
-      try {
-        log.info(`[BRAND-PRESENCE-FILTERS] Querying brand presence filter values for siteId: ${siteId}`);
-        const queryStart = Date.now();
-
-        // Query for distinct values of each filter dimension
-        // Using a single query with DISTINCT ON each column is more efficient than multiple queries
-        const [
-          categories,
-          topics,
-          models,
-          regions,
-          origins,
-        ] = await Promise.all([
-          // Get distinct categories
-          aurora.query(
-            `SELECT DISTINCT category 
-             FROM public.brand_presence 
-             WHERE category IS NOT NULL 
-               AND category != '' 
-             ORDER BY category`,
-          ),
-          // Get distinct topics (split comma-separated values)
-          aurora.query(
-            `SELECT DISTINCT TRIM(unnest(string_to_array(topics, ','))) as topic
-             FROM public.brand_presence 
-             WHERE topics IS NOT NULL 
-               AND topics != '' 
-             ORDER BY topic`,
-          ),
-          // Get distinct models (platforms)
-          aurora.query(
-            `SELECT DISTINCT model 
-             FROM public.brand_presence 
-             WHERE model IS NOT NULL 
-               AND model != '' 
-             ORDER BY model`,
-          ),
-          // Get distinct regions
-          aurora.query(
-            `SELECT DISTINCT region 
-             FROM public.brand_presence 
-             WHERE region IS NOT NULL 
-               AND region != '' 
-             ORDER BY region`,
-          ),
-          // Get distinct origins
-          aurora.query(
-            `SELECT DISTINCT origin 
-             FROM public.brand_presence 
-             WHERE origin IS NOT NULL 
-               AND origin != '' 
-             ORDER BY origin`,
-          ),
-        ]);
-
-        queryDuration = Date.now() - queryStart;
-
-        // Extract values from query results
-        filterValues = {
-          categories: categories.map((row) => row.category),
-          topics: topics.map((row) => row.topic),
-          models: models.map((row) => row.model),
-          regions: regions.map((row) => row.region),
-          origins: origins.map((row) => row.origin),
-        };
-
-        log.info(`[BRAND-PRESENCE-FILTERS] Filter values retrieved for siteId: ${siteId} - categories: ${filterValues.categories.length}, topics: ${filterValues.topics.length}, models: ${filterValues.models.length}, regions: ${filterValues.regions.length}, origins: ${filterValues.origins.length}`);
-      } catch (dbError) {
-        log.error(`[BRAND-PRESENCE-FILTERS] Database query failed for siteId: ${siteId} - error: ${dbError.message}`);
-        return badRequest(`Failed to fetch filter values: ${dbError.message}`);
-      }
-
-      const totalDuration = Date.now() - startTime;
-      log.info(`[BRAND-PRESENCE-FILTERS] Request completed for siteId: ${siteId} - total duration: ${totalDuration}ms`);
-
-      return ok({
-        siteId,
-        filters: filterValues,
-        performance: {
-          totalDuration,
-          queryDuration,
-          validationDuration: validationDuration || 0,
-        },
-      }, {
-        'access-control-allow-origin': '*',
-        'access-control-allow-methods': 'GET, OPTIONS',
-        'access-control-allow-headers': 'x-api-key, authorization, content-type',
-      });
-    } catch (error) {
-      const totalDuration = Date.now() - startTime;
-      log.error(`[BRAND-PRESENCE-FILTERS] Request failed for siteId: ${siteId} - duration: ${totalDuration}ms, error: ${error.message}`);
-      return badRequest(error.message);
-    }
-  };
+  // Wrapper for brand presence example endpoint
+  const exampleEndpoint = async (context) => exampleEndpointImpl(
+    context,
+    getSiteAndValidateLlmo,
+  );
 
   return {
     getLlmoSheetData,
@@ -1175,8 +979,8 @@ function LlmoController(ctx) {
     onboardCustomer,
     offboardCustomer,
     queryFiles,
-    getLlmoAurora,
     getBrandPresenceFilters,
+    exampleEndpoint,
   };
 }
